requirements.txt
pytest==7.4.2
requests==2.31.0
ipinfo==4.4.3
PyYAML==6.0.1